address feedback and rename component
