+++ src/components/footer-legal-strip/__tests__/footer-legal-strip-test-cases.js
@@ -1,4 +1,4 @@
-import PageFooter from '../page-footer';
+import PageFooter from '../footer-legal-strip';
 
 const testCases = {};
 testCases.basic = {
