+++ src/components/footer-legal-strip/index.js
@@ -1,0 +1,3 @@
+import main from './footer-legal-strip';
+
+export default main;
