--- src/components/page-footer/index.js
@@ -1,3 +1,0 @@
-import main from './page-footer';
-
-export default main;
